refactor(outcome): move skill discovery and bundle fingerprinting to outcome_discovery (#1214)
outcome_cmd.py is frozen at 2903 lines by the module-size ratchet and the
finding-4 anchored discovery pushed it to 2947. The registry enumeration
(_skill_names_at, _linked_worktree_skill_roots, _artifact_content_path_at)
and the snapshot fingerprint helpers (_bundle_fingerprint,
_bundle_fingerprint_from_files) now live in outcome_discovery.py;
artifact_fingerprint stays on outcome_cmd and delegates skill bundles.
Behaviour unchanged; outcome_cmd.py is 2751 lines.
